Add files via upload
Here is an example of the proposed methodology:
Step 1: DEM2G Toolbox (Nested Hierarchy of Depressions and Filling-Spilling-Merging Sequence)
Inputs:
•	Digital Elevation Model (DEM)
•	CN (Curve Number)
Outputs:
•	NETCDF file
•	Table
The first step involves preprocessing the DEM using the ArcGIS Python toolbox provided in this repository. This toolbox includes two main components: "Functions.py" and "DEM2G.pyt". By utilizing these tools, the DEM data is processed to establish the nested hierarchy of depressions and determine the sequence of filling, spilling, and merging.
Step 2: GB-RFSM Implementation
Inputs:
•	Output from Step 1 (NETCDF and table)
•	Road raster file
In the second step, a Python program called "RFSMtoolsV3.py" is employed to generate the DEM-graph and execute the GB-RFSM (Graph-Based Rainfall-Runoff Simulation Model). The input for this step includes the outputs from Step 1, which are the NETCDF file and table. Additionally, a road raster file is required to enhance the accuracy of the simulation.
By following this two-step methodology and utilizing the provided tools and programs, you can analyze digital elevation data, establish a nested hierarchy of depressions, and simulate rainfall-runoff processes using the GB-RFSM.